[refactor] move EnsureLocalGlobIgnored to fileutil
Addresses the open design-note thread: the function belongs in fileutil
alongside the other gitignore helpers rather than in trust. Added
config.TrustFile constant to break the potential import cycle
(fileutil cannot import trust).
